+++ src/Grappachu.Core.Test/Collections/CollectionUtilsTest.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Drawing;
 using System.Linq;
-using System.Security.Cryptography.X509Certificates;
 using Grappachu.Core.Collections;
 using NUnit.Framework;
 using SharpTestsEx;