+++ src/Grappachu.Core.Test/Security/Encryption/AesTest.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using Grappachu.Core.Lang;
 using Grappachu.Core.Lang.Text;
 using Grappachu.Core.Preview.Encryption;
-using Grappachu.Core.Security;
 using NUnit.Framework;
 using SharpTestsEx;