+++ src/Grappachu.Core/Lang/Extensions/StringExtensions.cs
@@ -140,5 +140,8 @@ namespace Grappachu.Core.Lang.Extensions
             }
             return sb.ToString();
         }
+
+
+
     }
 }